requirements.txt
#python_version>=3.6
requests>=2.19.1
PyYAML>=3.13
